Normalize unstable yaml-syntax-error
